Sets up new calibration to work with Tobii tracker.
